+++ linux/keyman-config/keyman_config/convertico.py
@@ -4,6 +4,7 @@ import logging
 import numpy as np
 import os
 import sys
+import struct
 from PIL import Image
 
 
@@ -19,6 +20,13 @@ def changeblacktowhite(im):
     return im2
 
 def checkandsaveico(icofile):
+	"""
+	Convert keyman ico file to png to work in IBus
+    The ico file may be ico or bmp format
+
+	Args:
+		icofile (str): path to ico file
+	"""
     im = Image.open(icofile)
     im = im.convert('RGBA')
     im2 = im
@@ -33,13 +41,47 @@ def checkandsaveico(icofile):
     im4.save(icofile + ".png")
     os.remove(icofile + ".bmp")
 
+def extractico(kmxfile):
+	"""
+	Extract icon file from compiled kmx keyboard
+
+	Args:
+		kmxfile (str): path to kmx file
+	"""
+    name, ext = os.path.splitext(kmxfile)
+    icofilename = name+".ico"
+    with open(kmxfile, mode='rb') as file: # b is important -> binary
+        fileContent = file.read()
+
+        kmxstart = struct.unpack_from("<16I", fileContent, 0)
+        if kmxstart[0] != 0x5354584B:
+            logging.debug("bad kmx identifier")
+            return False
+        bitmapOffset = kmxstart[14]
+        bitmapSize = kmxstart[15]
+        logging.debug("bitmap offset is %d", bitmapOffset)
+        logging.debug("bitmap size is %d", bitmapSize)
+        file.seek(bitmapOffset, 0)
+        bitmap = file.read(bitmapSize)
+        if not bitmap:
+            logging.debug("unreadable bitmap in kmx")
+            return False
+        with open(icofilename, mode='wb') as iconfile:
+            iconfile.write(bitmap)
+        checkandsaveico(icofilename)
+        return True
+
 
 def main(argv):
     if len(sys.argv) != 2:
-        logging.error("convertico.py <ico file>")
+        logging.error("convertico.py <ico file | kmx file>")
         sys.exit(2)
     logging.basicConfig(level=logging.INFO)
-    checkandsaveico(sys.argv[1])
+    name, ext = os.path.splitext(sys.argv[1])
+    if ext == ".kmx":
+        extractico(sys.argv[1])
+    else:
+        checkandsaveico(sys.argv[1])
 
 if __name__ == "__main__":
     main(sys.argv[1:])